Clean up and refactor socket support
