+++ common/src/iosMain/kotlin/fr/acinq/eklair/io/NativeSocket.kt
@@ -1,4 +1,4 @@
-package fr.acinq.eklair
+package fr.acinq.eklair.io
 
 /**
  * extracted and adapted from
@@ -16,9 +16,13 @@ class NativeSocket private constructor(internal val sockfd: Int, private var end
             init_sockets()
         }
 
-        operator fun invoke() = NativeSocket(socket(AF_INET, SOCK_STREAM, 0), Endpoint(IP(0, 0, 0, 0), 0))
-        suspend fun connect(host: String, port: Int) = NativeSocket().apply { connect(host, port) }
-        suspend fun bound(host: String, port: Int) = NativeSocket().apply { bind(host, port) }
+        operator fun invoke() = NativeSocket(
+            socket(AF_INET, SOCK_STREAM, 0),
+            Endpoint(IP(0, 0, 0, 0), 0))
+        suspend fun connect(host: String, port: Int) = NativeSocket()
+            .apply { connect(host, port) }
+        suspend fun bound(host: String, port: Int) = NativeSocket()
+            .apply { bind(host, port) }
     }
 
     data class Endpoint(val ip: IP, val port: Int) {
@@ -27,12 +31,12 @@ class NativeSocket private constructor(internal val sockfd: Int, private var end
 
     class IP(val data: UByteArray) {
         constructor(v0: Int, v1: Int, v2: Int, v3: Int) : this(
-            ubyteArrayOf(
-                v0.toUByte(),
-                v1.toUByte(),
-                v2.toUByte(),
-                v3.toUByte()
-            )
+                ubyteArrayOf(
+                        v0.toUByte(),
+                        v1.toUByte(),
+                        v2.toUByte(),
+                        v3.toUByte()
+                )
         )
 
         val v0 get() = data[0]
@@ -98,8 +102,8 @@ class NativeSocket private constructor(internal val sockfd: Int, private var end
 
     fun CPointer<sockaddr_in>.toEndpoint(): Endpoint {
         return Endpoint(
-            IP(this.pointed.sin_addr.readValue().getBytes().toUByteArray()),
-            swapBytes(this.pointed.sin_port.toUShort()).toInt()
+                IP(this.pointed.sin_addr.readValue().getBytes().toUByteArray()),
+                swapBytes(this.pointed.sin_port.toUShort()).toInt()
         )
     }
 
@@ -118,7 +122,10 @@ class NativeSocket private constructor(internal val sockfd: Int, private var end
                 }
             }
             //println("accept: fd=$fd")
-            return NativeSocket(fd, addr.ptr.reinterpret<sockaddr_in>().toEndpoint()).apply {
+            return NativeSocket(
+                fd,
+                addr.ptr.reinterpret<sockaddr_in>().toEndpoint()
+            ).apply {
                 setSocketBlockingEnabled(false)
             }
         }
@@ -202,14 +209,18 @@ class NativeSocket private constructor(internal val sockfd: Int, private var end
             //println("result: $result")
             //println("local address: " + inet_ntoa(localAddress.sin_addr.readValue())?.toKString())
             //println("local port: " + )
-            return Endpoint(IP(ip.getBytes().toUByteArray()), port.toInt())
+            return Endpoint(
+                IP(
+                    ip.getBytes().toUByteArray()
+                ), port.toInt()
+            )
         }
     }
 
     fun getRemoteEndpoint() = endpoint
 
     private fun swapBytes(v: UShort): UShort =
-        (((v.toInt() and 0xFF) shl 8) or ((v.toInt() ushr 8) and 0xFF)).toUShort()
+            (((v.toInt() and 0xFF) shl 8) or ((v.toInt() ushr 8) and 0xFF)).toUShort()
 
     override fun toString(): String = "NativeSocket(local=${getLocalEndpoint()}, remote=${getRemoteEndpoint()})"
 }
